code simplification
I have been trying to push as much code as possible down to C, and was
trying to implement encoding with padding also in C. But I am not liking
the direction this is taking: the C code is no longer as easy to follow,
and it is deviating from the reference implementation.

I prefer the C functions to be fundamentally equal to the ones in the
reference implementation, and to wrap them in a bit of Ruby.